requirements.txt
keras_nightly==2.5.0.dev2021032900
torch==1.8.1
tensorflow==2.5.0
opencv_python==4.5.2.54
numpy==1.19.5
scikit_image==0.18.1
matplotlib==3.4.2
torchvision==0.9.1
pandas==1.2.4
keras==2.4.3
Pillow==8.2.0
scikit_learn==0.24.2
skimage==0.0
